add user managemant related pages

+++ src/services/userService.js
@@ -3,19 +3,19 @@ import api from './api';
 const userService = {
     // Admin: Get all users
     getAllUsers: async () => {
-        const response = await api.get('/users');
+        const response = await api.get('/admin/users');
         return response.data;
     },
 
     // Admin: Update role
     updateUserRole: async (id, role) => {
-        const response = await api.patch(`/users/${id}/role`, { role });
+        const response = await api.patch(`/admin/users/${id}/role`, { role });
         return response.data;
     },
 
     // Admin: Delete user
     deleteUser: async (id) => {
-        const response = await api.delete(`/users/${id}`);
+        const response = await api.delete(`/admin/users/${id}`);
         return response.data;
     }
 };
